Refactor ecosystem stats into managers
